Remove budget configuration fields (#120)
Remove the budget feature (BudgetConfig, Cadence, budget DB table)
per maintainer decision. DonationRecord moved into storage module
where it is exclusively used. Donation history tracking is preserved.
