+++ src/lib.rs
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-pub mod budget;
 pub mod config;
 pub mod contribute;
 pub mod discover;
